+++ doc/release_notes.rst
@@ -75,6 +75,8 @@ Upcoming Release
 
 * Revise OSM cleaning to improve the cleaning process and error resilience `PR #620 https://github.com/pypsa-meets-earth/pypsa-earth/pull/620`__
 
+* Fix isolated buses when simplifying the network and add clustering by networks `PR #632 https://github.com/pypsa-meets-earth/pypsa-earth/pull/632`__
+
 PyPSA-Earth 0.1.0
 =================